* [e2e] Improve tests for single node environment
